First Iteration
In this first iteration of the SEPIA filter I made a function to make the integer value results into a BYTE. I did this because the math operations done in the SEPIA function results are sometimes a number smaller than 0 and bigger than 255. 0 - 255 is an inclusive range and important because, for the function to return the data type of BYTE, an integer's value must remain with the inclusive range. 

This iteration was not a success. 

First, I made a copy of the original image declared in line 32 RGBTRIPLE rgbt . T
Second, I took the initial RGB values of the image at coordinates [i][j]  to alter them
Third, I stored the altered copy in the original pixel at coordinate [i][j]

The problem with this was that I kept getting a boxy image in pink.  I was getting a boxy image because. I don't know why. But in the next iteration, when I did not use a copy and made changes to the original image[i][j], the image printed perfectly in SEPIA.