Second Iteration
This iteration successfully changed the original image to SEPIA. This time I did not make a copy of the image[i][j] as and RGBTRIPLE. Instead, I worked directly with the pixels in the image at coordinate [i][j]. 

I think the reason this worked as opposed to the copy was because the conversion of data from copy was a little messy in operations and in this iteration I worked directly with the data I wanted to manipulate. My hypothesis is that when I worked with a copy to alter and then putting it back into the original must truncate some data. This truncation may alter the image. I will test this hypothesis.

In any case, I find this iteration of code more elegant.