Blur Image Addition
This is the third iteration of a code for the image filter. Here I added the "blur" feature. I have figured out how to alter the Red, Green, and Blue values for the neighboring pixels, however, due to a conditional that I needed to apply for each mathematical operation I could not alter the pixels. I need conditionals for specific pixels such as those who do not have a full set of 9 neighbors.